requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==1.2.0
    # via
    #   tensorboard
    #   tensorflow
aiohttp==3.8.1
    # via
    #   aiohttp-cors
    #   bentoml
    #   faust-streaming
aiohttp-cors==0.7.0
    # via faust-streaming
aiokafka==0.7.2
    # via faust-streaming
aiosignal==1.2.0
    # via aiohttp
anyio==3.6.1
    # via
    #   starlette
    #   watchfiles
appdirs==1.4.4
    # via fs
asgiref==3.5.2
    # via opentelemetry-instrumentation-asgi
astunparse==1.6.3
    # via tensorflow
async-timeout==4.0.2
    # via aiohttp
asynctest==0.13.0
    # via aiohttp
attrs==22.1.0
    # via
    #   aiohttp
    #   bentoml
    #   cattrs
backports-cached-property==1.0.2
    # via bentoml
bentoml==1.0.4
    # via -r requirements.in
build==0.8.0
    # via pip-tools
cachetools==5.2.0
    # via google-auth
cattrs==22.1.0
    # via bentoml
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via
    #   aiohttp
    #   requests
circus==0.17.1
    # via bentoml
click==8.0.4
    # via
    #   bentoml
    #   faust-streaming
    #   nltk
    #   pip-tools
    #   uvicorn
cloudpickle==2.1.0
    # via bentoml
colorlog==6.6.0
    # via mode-streaming
commonmark==0.9.1
    # via rich
contextlib2==21.6.0
    # via schema
croniter==1.3.5
    # via
    #   faust-streaming
    #   mode-streaming
deepmerge==1.0.1
    # via bentoml
deprecated==1.2.13
    # via opentelemetry-api
exceptiongroup==1.0.0rc8
    # via cattrs
faust-streaming==0.8.8
    # via -r requirements.in
flatbuffers==1.12
    # via tensorflow
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
fs==2.4.16
    # via bentoml
gast==0.4.0
    # via tensorflow
google-auth==2.11.0
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.47.0
    # via
    #   tensorboard
    #   tensorflow
h11==0.13.0
    # via uvicorn
h5py==3.7.0
    # via tensorflow
idna==3.3
    # via
    #   anyio
    #   requests
    #   yarl
importlib-metadata==4.12.0
    # via
    #   bentoml
    #   build
    #   click
    #   markdown
    #   pep517
intervaltree==3.1.0
    # via faust-streaming
jinja2==3.1.2
    # via bentoml
joblib==1.1.0
    # via
    #   nltk
    #   scikit-learn
kafka-python==2.0.2
    # via aiokafka
keras==2.9.0
    # via tensorflow
keras-preprocessing==1.1.2
    # via tensorflow
libclang==14.0.6
    # via tensorflow
markdown==3.4.1
    # via tensorboard
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
mode-streaming==0.2.0
    # via faust-streaming
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via
    #   faust-streaming
    #   mode-streaming
nltk==3.7
    # via -r requirements.in
numpy==1.21.6
    # via
    #   bentoml
    #   h5py
    #   keras-preprocessing
    #   opt-einsum
    #   pandas
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   tensorflow
oauthlib==3.2.0
    # via requests-oauthlib
opentelemetry-api==1.12.0
    # via
    #   bentoml
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-sdk
opentelemetry-instrumentation==0.33b0
    # via
    #   bentoml
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-aiohttp-client==0.33b0
    # via bentoml
opentelemetry-instrumentation-asgi==0.33b0
    # via bentoml
opentelemetry-sdk==1.12.0
    # via bentoml
opentelemetry-semantic-conventions==0.33b0
    # via
    #   bentoml
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-sdk
opentelemetry-util-http==0.33b0
    # via
    #   bentoml
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asgi
opentracing==2.4.0
    # via faust-streaming
opt-einsum==3.3.0
    # via tensorflow
packaging==21.3
    # via
    #   bentoml
    #   build
    #   tensorflow
pandas==1.3.5
    # via -r requirements.in
pathspec==0.9.0
    # via bentoml
pep517==0.13.0
    # via build
pip-tools==6.8.0
    # via bentoml
prometheus-client==0.13.1
    # via bentoml
protobuf==3.19.4
    # via
    #   tensorboard
    #   tensorflow
psutil==5.9.1
    # via
    #   bentoml
    #   circus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pygments==2.13.0
    # via rich
pynvml==11.4.1
    # via bentoml
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via
    #   bentoml
    #   croniter
    #   pandas
python-dotenv==0.20.0
    # via bentoml
python-multipart==0.0.5
    # via bentoml
pytz==2022.2.1
    # via pandas
pyyaml==6.0
    # via bentoml
pyzmq==23.2.1
    # via circus
regex==2022.8.17
    # via nltk
requests==2.28.1
    # via
    #   bentoml
    #   requests-oauthlib
    #   tensorboard
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rich==12.5.1
    # via bentoml
rsa==4.9
    # via google-auth
schema==0.7.5
    # via bentoml
scikit-learn==1.0.2
    # via -r requirements.in
scipy==1.7.3
    # via scikit-learn
simple-di==0.1.5
    # via bentoml
six==1.16.0
    # via
    #   astunparse
    #   fs
    #   google-auth
    #   google-pasta
    #   grpcio
    #   keras-preprocessing
    #   python-dateutil
    #   python-multipart
    #   tensorflow
sniffio==1.2.0
    # via anyio
sortedcontainers==2.4.0
    # via intervaltree
starlette==0.20.4
    # via bentoml
tensorboard==2.9.1
    # via tensorflow
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
tensorflow==2.9.1
    # via -r requirements.in
tensorflow-estimator==2.9.0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.26.0
    # via tensorflow
termcolor==1.1.0
    # via tensorflow
terminaltables==3.1.10
    # via faust-streaming
threadpoolctl==3.1.0
    # via scikit-learn
tomli==2.0.1
    # via
    #   build
    #   pep517
tornado==6.2
    # via circus
tqdm==4.64.0
    # via nltk
typing-extensions==4.3.0
    # via
    #   aiohttp
    #   anyio
    #   asgiref
    #   async-timeout
    #   cattrs
    #   h11
    #   importlib-metadata
    #   mode-streaming
    #   opentelemetry-sdk
    #   rich
    #   starlette
    #   tensorflow
    #   uvicorn
    #   yarl
urllib3==1.26.12
    # via requests
uvicorn==0.18.3
    # via bentoml
venusian==3.0.0
    # via faust-streaming
watchfiles==0.16.1
    # via bentoml
werkzeug==2.2.2
    # via tensorboard
wheel==0.37.1
    # via
    #   astunparse
    #   pip-tools
    #   tensorboard
wrapt==1.14.1
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
    #   tensorflow
yarl==1.8.1
    # via
    #   aiohttp
    #   faust-streaming
zipp==3.8.1
    # via
    #   importlib-metadata
    #   pep517

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
